Split __main__.py god module into cli/ package
The 1851-line __main__.py contained argparse, all 9 subcommand handlers,
ANSI terminal control, LLM patch management, git remote probing, and a
config-loading preamble copy-pasted 7 times with accidental variations.

Extract into a cli/ package:
- cli/__init__.py: dispatcher (main, claudelint_shim)
- cli/_config.py: shared load_config() replacing 7 copies
- cli/_parser.py: argparse tree (~400 lines)
- cli/_helpers.py: shared utilities (progress, paths, colors, patches, LLM)
- cli/_lint.py, _fix.py, _badge.py, _explain.py, _docs.py, _tree.py,
  _baseline.py, _simple.py: one module per handler

__main__.py becomes a 5-line re-export. Fixes docs missing -c flag.
