+++ src/skillsaw/__main__.py
@@ -1,1851 +1,6 @@
-"""
-Entry point for skillsaw (and claudelint backward-compat shim)
-"""
+"""Entry point for skillsaw (and claudelint backward-compat shim)"""
 
-import argparse
-import logging
-import os
-import shutil
-import sys
-import time
-from pathlib import Path
-from importlib.metadata import version, PackageNotFoundError
-
-from .context import RepositoryContext, RepositoryType
-from .config import LinterConfig, find_config
-from .linter import Linter
-from .rule import AutofixConfidence, AutofixResult, Severity
-from .formatters import (
-    format_report,
-    get_counts,
-    infer_format,
-    parse_output_spec,
-    FORMATS,
-    EXTENSION_MAP,
-)
-from . import __version__
-
-_SUBCOMMANDS = {
-    "lint",
-    "init",
-    "list-rules",
-    "docs",
-    "add",
-    "fix",
-    "tree",
-    "baseline",
-    "explain",
-    "badge",
-}
-
-
-def _get_version() -> str:
-    try:
-        return version("skillsaw")
-    except PackageNotFoundError:
-        return __version__
-
-
-def _emit_config_warnings(config) -> None:
-    """Print non-fatal config-load warnings (missing version, unknown keys)."""
-    for warning in getattr(config, "warnings", []):
-        print(f"Warning: {warning}", file=sys.stderr)
-
-
-def _build_parser():
-    """Build the main argument parser with all subcommands.
-
-    Extracted so that documentation generators can introspect the real parser
-    without running main().
-    """
-    parser = argparse.ArgumentParser(
-        prog="skillsaw",
-        description="Keep your skills sharp.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  skillsaw                        # Lint current directory
-  skillsaw lint /path/to/skills   # Lint specific directory
-  skillsaw init                   # Generate default config
-  skillsaw list-rules             # List available rules
-  skillsaw docs                   # Generate documentation
-  skillsaw add marketplace        # Scaffold a new marketplace
-
-For more information, visit: https://github.com/stbenjam/skillsaw
-        """,
-    )
-    parser.add_argument("--version", action="version", version=f"%(prog)s {_get_version()}")
-
-    subparsers = parser.add_subparsers(dest="command")
-
-    # --- lint ---
-    lint_parser = subparsers.add_parser(
-        "lint",
-        help="Lint agent skills, plugins, and AI coding assistant context",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    lint_parser.add_argument(
-        "path",
-        nargs="*",
-        type=Path,
-        default=[Path.cwd()],
-        help="Paths to skill, plugin, or marketplace directories/files (default: current directory)",
-    )
-    lint_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file (default: auto-discover from the first path)",
-    )
-    lint_parser.add_argument(
-        "-v", "--verbose", action="store_true", help="Show info-level messages"
-    )
-    lint_parser.add_argument(
-        "--strict",
-        action="store_true",
-        help="Treat warnings as errors (exit with error code if warnings exist)",
-    )
-    # Deprecated: use `skillsaw fix` instead. Hidden from --help.
-    lint_parser.add_argument("--fix", action="store_true", help=argparse.SUPPRESS)
-    lint_parser.add_argument(
-        "--llm", "--ai", action="store_true", dest="use_llm", help=argparse.SUPPRESS
-    )
-    lint_parser.add_argument(
-        "--dry-run", action="store_true", dest="dry_run", help=argparse.SUPPRESS
-    )
-    lint_parser.add_argument(
-        "--patch-file", type=Path, default=None, dest="patch_file", help=argparse.SUPPRESS
-    )
-    lint_parser.add_argument(
-        "--apply-patch", action="store_true", dest="apply_patch", help=argparse.SUPPRESS
-    )
-    lint_parser.add_argument(
-        "--format",
-        dest="fmt",
-        default="text",
-        choices=FORMATS,
-        help="Output format for stdout (default: text)",
-    )
-    lint_parser.add_argument(
-        "--output",
-        dest="outputs",
-        action="append",
-        default=[],
-        metavar="[FORMAT:]FILE",
-        help="Write output to FILE. Format is inferred from extension "
-        f"({', '.join(sorted(EXTENSION_MAP))}) "
-        "or set explicitly with a FORMAT: prefix (e.g. gitlab:report.json). "
-        "Use the prefix when an extension is ambiguous (e.g. .json could be "
-        "json or gitlab/code-climate). Can be specified multiple times.",
-    )
-    lint_parser.add_argument(
-        "--type",
-        dest="repo_types",
-        action="append",
-        default=[],
-        metavar="TYPE",
-        help="Override auto-detected repository type (repeatable). "
-        "Values: "
-        + ", ".join(t.value for t in RepositoryType if t is not RepositoryType.UNKNOWN)
-        + ".",
-    )
-    lint_parser.add_argument(
-        "--rule",
-        dest="rule_ids",
-        action="append",
-        default=[],
-        metavar="RULE",
-        help="Only run these rules (repeatable). Config still comes from .skillsaw.yaml.",
-    )
-    lint_parser.add_argument(
-        "--skip-rule",
-        dest="skip_rule_ids",
-        action="append",
-        default=[],
-        metavar="RULE",
-        help="Skip these rules (repeatable). Cannot be combined with --rule.",
-    )
-    lint_parser.add_argument(
-        "--no-baseline",
-        action="store_true",
-        dest="no_baseline",
-        help="Ignore baseline file even if .skillsaw-baseline.json exists",
-    )
-    lint_parser.add_argument(
-        "--no-custom-rules",
-        action="store_true",
-        dest="no_custom_rules",
-        help="Skip custom rules defined in .skillsaw.yaml (recommended for CI on untrusted PRs)",
-    )
-    lint_parser.add_argument(
-        "--no-progress",
-        action="store_true",
-        dest="no_progress",
-        help="Disable the interactive per-rule progress indicator "
-        "(auto-disabled when stderr is not a terminal)",
-    )
-
-    # --- fix ---
-    fix_parser = subparsers.add_parser(
-        "fix",
-        help="Automatically fix lint violations",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    fix_parser.add_argument(
-        "path",
-        nargs="*",
-        type=Path,
-        default=[Path.cwd()],
-        help="Paths to repositories or files (default: current directory)",
-    )
-    fix_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file (default: auto-discover from the first path)",
-    )
-    fix_parser.add_argument(
-        "--llm",
-        "--ai",
-        action="store_true",
-        dest="use_llm",
-        help="Use LLM-powered fixes for content violations",
-    )
-    fix_parser.add_argument(
-        "--model",
-        help="Override LLM model (default: from config or claude-sonnet-4-20250514)",
-    )
-    fix_parser.add_argument(
-        "--max-iterations",
-        type=int,
-        help="Max fix iterations per file (default: 3)",
-    )
-    fix_parser.add_argument(
-        "--all",
-        action="store_true",
-        help="Include info-level violations (default: only errors and warnings)",
-    )
-    fix_parser.add_argument(
-        "-y",
-        "--yes",
-        action="store_true",
-        help="Auto-apply changes without confirmation",
-    )
-    fix_parser.add_argument(
-        "--workers",
-        type=int,
-        default=None,
-        help="Number of parallel LLM workers (default: 4)",
-    )
-    fix_parser.add_argument(
-        "--dry-run",
-        action="store_true",
-        dest="dry_run",
-        help="Preview fixes without writing changes",
-    )
-    fix_parser.add_argument(
-        "--patch-file",
-        type=Path,
-        default=None,
-        dest="patch_file",
-        metavar="FILE",
-        help="Path for the saved LLM patch file (default: .skillsaw-llm-patch.diff in repo root)",
-    )
-    fix_parser.add_argument(
-        "--apply-patch",
-        action="store_true",
-        dest="apply_patch",
-        help="Apply a previously saved LLM dry-run patch (use --patch-file to specify path)",
-    )
-    fix_parser.add_argument(
-        "--suggest",
-        action="store_true",
-        help="Also apply suggested fixes (not just safe ones)",
-    )
-    fix_parser.add_argument(
-        "--rule",
-        dest="rule_ids",
-        action="append",
-        default=[],
-        metavar="RULE",
-        help="Only run these rules (repeatable). Config still comes from .skillsaw.yaml.",
-    )
-    fix_parser.add_argument(
-        "--skip-rule",
-        dest="skip_rule_ids",
-        action="append",
-        default=[],
-        metavar="RULE",
-        help="Skip these rules (repeatable). Cannot be combined with --rule.",
-    )
-    fix_parser.add_argument(
-        "--no-custom-rules",
-        action="store_true",
-        dest="no_custom_rules",
-        help="Skip custom rules defined in .skillsaw.yaml (recommended for CI on untrusted PRs)",
-    )
-    fix_parser.add_argument(
-        "--no-progress",
-        action="store_true",
-        dest="no_progress",
-        help="Disable the interactive per-rule progress indicator "
-        "(auto-disabled when stderr is not a terminal)",
-    )
-
-    # --- init ---
-    init_parser = subparsers.add_parser(
-        "init", help="Generate a default .skillsaw.yaml config file"
-    )
-    init_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Directory to create config in (default: current directory)",
-    )
-
-    # --- list-rules ---
-    subparsers.add_parser("list-rules", help="List all available builtin rules")
-
-    # --- explain ---
-    explain_parser = subparsers.add_parser(
-        "explain",
-        help="Show documentation and effective configuration for a rule",
-    )
-    explain_parser.add_argument(
-        "rule_id",
-        metavar="RULE",
-        help="Rule ID to explain (e.g. content-weak-language)",
-    )
-    explain_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Repository to compute effective config in (default: current directory)",
-    )
-    explain_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file (default: auto-discover)",
-    )
-
-    # --- docs ---
-    docs_parser = subparsers.add_parser(
-        "docs",
-        help="Generate documentation for a plugin, marketplace, or .claude repository",
-        description="Generate documentation for a plugin, marketplace, or .claude repository",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    docs_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Path to repository (default: current directory)",
-    )
-    docs_parser.add_argument(
-        "--format",
-        dest="fmt",
-        default="html",
-        choices=["html", "markdown"],
-        help="Output format (default: html)",
-    )
-    docs_parser.add_argument(
-        "-o",
-        "--output",
-        type=Path,
-        default=None,
-        help="Output file or directory (default: skillsaw-docs/). "
-        "If it ends with .html/.md, writes a single file directly.",
-    )
-    docs_parser.add_argument("--title", default=None, help="Custom title for the documentation")
-    docs_parser.add_argument(
-        "--theme",
-        default=None,
-        help="Color theme for HTML output. Presets: indigo (default), forest-green, "
-        "ocean-blue, sunset-orange, royal-purple, crimson-red.",
-    )
-
-    # --- tree ---
-    tree_parser = subparsers.add_parser(
-        "tree",
-        help="Display the repository lint tree",
-    )
-    tree_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Path to repository (default: current directory)",
-    )
-    tree_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file",
-    )
-    tree_parser.add_argument(
-        "--format",
-        dest="fmt",
-        default="text",
-        choices=["text", "dot"],
-        help="Output format (default: text)",
-    )
-
-    # --- baseline ---
-    baseline_parser = subparsers.add_parser(
-        "baseline",
-        help="Generate or update the baseline file from current violations",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    baseline_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Path to repository (default: current directory)",
-    )
-    baseline_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file",
-    )
-    # --- badge ---
-    badge_parser = subparsers.add_parser(
-        "badge",
-        help="Grade the repository and write a shields.io badge JSON file",
-        description="Lint the repository, compute its letter grade, write a "
-        "shields.io-compatible badge file, and print the markdown to embed it. "
-        "Ignores any baseline so the published grade reflects all violations.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    badge_parser.add_argument(
-        "path",
-        nargs="?",
-        type=Path,
-        default=Path.cwd(),
-        help="Path to repository (default: current directory)",
-    )
-    badge_parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        help="Path to .skillsaw.yaml config file (default: auto-discover)",
-    )
-    badge_parser.add_argument(
-        "-o",
-        "--output",
-        type=Path,
-        default=None,
-        help="Badge JSON output path (default: .skillsaw-badge.json in the repository root)",
-    )
-
-    # --- add ---
-    subparsers.add_parser(
-        "add",
-        help="Scaffold marketplaces, plugins, skills, commands, agents, and hooks",
-        add_help=False,
-    )
-
-    return parser
-
-
-def main():
-    # When no subcommand is given (or the first arg looks like a path/flag),
-    # default to "lint" so bare `skillsaw` and `skillsaw /path` keep working.
-    # `add` has its own argparse — dispatch before the main parser sees the args.
-    if len(sys.argv) > 1 and sys.argv[1] == "add":
-        from .marketplace.cli import _run_add_cli
-
-        return _run_add_cli()
-
-    if len(sys.argv) < 2 or sys.argv[1] not in _SUBCOMMANDS | {"--version", "-h", "--help"}:
-        sys.argv.insert(1, "lint")
-
-    parser = _build_parser()
-    args = parser.parse_args()
-
-    if args.command == "lint":
-        _run_lint(args)
-    elif args.command == "fix":
-        _run_fix(args)
-    elif args.command == "init":
-        _run_init(args)
-    elif args.command == "list-rules":
-        _run_list_rules()
-    elif args.command == "explain":
-        _run_explain(args)
-    elif args.command == "docs":
-        _run_docs(args)
-    elif args.command == "baseline":
-        _run_baseline(args)
-    elif args.command == "badge":
-        _run_badge(args)
-    elif args.command == "tree":
-        _run_tree(args)
-
-
-def _run_tree(args):
-    if not args.path.exists():
-        print(f"Error: Path not found: {args.path}", file=sys.stderr)
-        sys.exit(1)
-
-    context = RepositoryContext(args.path)
-
-    if args.config:
-        config_path = args.config
-        if not config_path.exists():
-            print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        config_path = find_config(args.path)
-
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-
-    context.content_paths = config.content_paths
-    context.exclude_patterns = config.exclude_patterns
-    context.apply_excludes()
-
-    tree = context.lint_tree
-    if args.fmt == "dot":
-        print(tree.print_dot(root_path=context.root_path))
-    else:
-        print(tree.print_tree(root_path=context.root_path))
-    sys.exit(0)
-
-
-class _RuleProgress:
-    """Single-line per-rule progress on stderr for interactive runs.
-
-    Inactive unless stderr is a terminal, so JSON/SARIF stdout, shell
-    pipelines, and CI logs are never polluted. Verbose mode also disables
-    it — info-level log lines share stderr and would interleave.
-    """
-
-    def __init__(self, args) -> None:
-        self.enabled = (
-            sys.stderr.isatty()
-            and not getattr(args, "no_progress", False)
-            and not getattr(args, "verbose", False)
-        )
-
-    def __call__(self, index: int, total: int, rule_id: str) -> None:
-        if self.enabled:
-            sys.stderr.write(f"\r\x1b[K  linting [{index}/{total}] {rule_id}")
-            sys.stderr.flush()
-
-    def clear(self) -> None:
-        """Erase the progress line before real output is printed."""
-        if self.enabled:
-            sys.stderr.write("\r\x1b[K")
-            sys.stderr.flush()
-
-
-def _resolve_lint_paths(paths):
-    """Normalize CLI paths into a unique list of directories to lint.
-
-    Files resolve to their parent directory, then exact duplicates and
-    paths nested inside another entry are dropped (a parent's
-    RepositoryContext already discovers everything beneath it).
-    First-seen order is preserved.
-    """
-    normalized = []
-    for p in paths:
-        resolved = p.resolve()
-        if resolved.is_file():
-            resolved = resolved.parent
-        normalized.append(resolved)
-
-    seen = set()
-    result = []
-    for p in normalized:
-        if p in seen:
-            continue
-        if any(p != other and _is_subpath(p, other) for other in normalized):
-            continue
-        seen.add(p)
-        result.append(p)
-    return result
-
-
-def _is_subpath(child, parent):
-    """Check if child is a strict subpath of parent."""
-    try:
-        child.relative_to(parent)
-        return child != parent
-    except ValueError:
-        return False
-
-
-class _MergedContext:
-    """Duck-typed context for formatters when linting multiple paths."""
-
-    def __init__(self, root_path, repo_types, plugins, skills):
-        self.root_path = root_path
-        self.repo_types = repo_types
-        self.plugins = plugins
-        self.skills = skills
-
-    @property
-    def repo_type(self):
-        for t in RepositoryContext._TYPE_PRIORITY:
-            if t in self.repo_types:
-                return t
-        return RepositoryType.UNKNOWN
-
-
-def _build_merged_context(contexts):
-    """Build a merged context from multiple RepositoryContexts."""
-    if len(contexts) == 1:
-        return contexts[0]
-    try:
-        root_path = Path(os.path.commonpath([c.root_path for c in contexts]))
-    except ValueError:
-        # No common ancestor (e.g. different drives on Windows)
-        root_path = contexts[0].root_path
-    repo_types = set()
-    plugins = []
-    skills = []
-    for ctx in contexts:
-        repo_types |= ctx.repo_types
-        plugins.extend(ctx.plugins)
-        skills.extend(ctx.skills)
-    return _MergedContext(root_path, repo_types, plugins, skills)
-
-
-def _dedup_rules(rules):
-    """Deduplicate rules by rule_id, preserving first occurrence."""
-    seen = set()
-    result = []
-    for rule in rules:
-        if rule.rule_id not in seen:
-            seen.add(rule.rule_id)
-            result.append(rule)
-    return result
-
-
-def _handle_apply_patch_for_lint(args):
-    if not getattr(args, "apply_patch", False):
-        return
-    paths = _resolve_lint_paths(args.path)
-    if not paths:
-        print("Error: No path provided for --apply-patch", file=sys.stderr)
-        sys.exit(1)
-    if len(paths) > 1:
-        print("Error: --apply-patch accepts a single path", file=sys.stderr)
-        sys.exit(1)
-    if not paths[0].exists():
-        print(f"Error: Path not found: {paths[0]}", file=sys.stderr)
-        sys.exit(1)
-    root_path = paths[0].resolve()
-    patch_path = _resolve_patch_path(args, root_path)
-    _apply_llm_patch(patch_path, root_path)
-    sys.exit(0)
-
-
-def _run_lint(args):
-    if args.verbose:
-        logging.basicConfig(
-            level=logging.INFO,
-            format="%(name)s: %(message)s",
-            stream=sys.stderr,
-        )
-    else:
-        logging.basicConfig(
-            level=logging.WARNING,
-            format="%(name)s: %(message)s",
-            stream=sys.stderr,
-        )
-
-    cli_version = _get_version()
-
-    output_formats = {}
-    for spec in args.outputs:
-        try:
-            fmt, filepath = parse_output_spec(spec)
-        except ValueError as e:
-            print(f"Error: {e}", file=sys.stderr)
-            sys.exit(1)
-        # Key on the resolved path so 'report.json' and './report.json' are
-        # recognised as the same target.
-        resolved_path = Path(filepath).resolve()
-        if resolved_path in output_formats and output_formats[resolved_path] != fmt:
-            print(
-                f"Error: --output targets '{filepath}' with conflicting formats "
-                f"'{output_formats[resolved_path]}' and '{fmt}'",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        output_formats[resolved_path] = fmt
-
-    _handle_apply_patch_for_lint(args)
-
-    # Validate paths as given, before file→parent normalization, so error
-    # messages reference exactly what the user typed
-    missing_count = 0
-    valid_raw_paths = []
-    for p in args.path:
-        if not p.exists():
-            print(f"Error: Path not found: {p}", file=sys.stderr)
-            missing_count += 1
-        else:
-            valid_raw_paths.append(p)
-
-    if missing_count:
-        print(f"{missing_count} path(s) not found", file=sys.stderr)
-
-    if not valid_raw_paths:
-        sys.exit(1)
-
-    paths = _resolve_lint_paths(valid_raw_paths)
-
-    if args.fmt == "text":
-        print(f"skillsaw {cli_version}")
-        print(f"Linting: {', '.join(str(p) for p in paths)}\n")
-
-    # Override repo_types if --type was provided
-    override_types = None
-    if args.repo_types:
-        type_map = {t.value: t for t in RepositoryType if t is not RepositoryType.UNKNOWN}
-        override_types = set()
-        for val in args.repo_types:
-            if val not in type_map:
-                print(
-                    f"Error: Unknown repository type '{val}'. "
-                    f"Valid types: {', '.join(sorted(type_map.keys()))}",
-                    file=sys.stderr,
-                )
-                sys.exit(1)
-            override_types.add(type_map[val])
-
-    if args.config:
-        config_path = args.config
-        if not config_path.exists():
-            print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        config_path = find_config(paths[0])
-
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-            if args.verbose and args.fmt == "text":
-                print(f"Using config: {config_path}\n")
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-
-    if args.strict:
-        config.strict = True
-
-    baseline = None
-    if not args.no_baseline:
-        from .baseline import find_baseline, load_baseline
-
-        baseline_path = find_baseline(config.config_dir or paths[0])
-
-        if baseline_path:
-            try:
-                baseline = load_baseline(baseline_path)
-                if args.verbose and args.fmt == "text":
-                    print(
-                        f"Using baseline: {baseline_path}"
-                        f" ({len(baseline.violations)} entries)\n"
-                    )
-            except (ValueError, OSError) as e:
-                print(f"Warning: Failed to load baseline: {e}", file=sys.stderr)
-
-    rule_ids = set(args.rule_ids) if args.rule_ids else None
-    skip_rule_ids = set(args.skip_rule_ids) if args.skip_rule_ids else None
-    if rule_ids and skip_rule_ids:
-        print("Error: --rule and --skip-rule cannot be combined", file=sys.stderr)
-        sys.exit(1)
-
-    # Create a context and linter per path, collect violations
-    lint_started = time.perf_counter()
-    all_violations = []
-    all_rules = []
-    contexts = []
-    baseline_suppressed = 0
-    for lint_path in paths:
-        context = RepositoryContext(lint_path, repo_types=override_types)
-        contexts.append(context)
-
-        if context.repo_type == RepositoryType.UNKNOWN:
-            print(
-                "Warning: Directory doesn't appear to be a recognized repository",
-                file=sys.stderr,
-            )
-            print(
-                "Expected: .claude-plugin/plugin.json, plugins/ directory,"
-                " or SKILL.md (agentskills.io)\n",
-                file=sys.stderr,
-            )
-
-        try:
-            linter = Linter(
-                context,
-                config,
-                rule_ids=rule_ids,
-                skip_rule_ids=skip_rule_ids,
-                baseline=baseline,
-                no_custom_rules=args.no_custom_rules,
-            )
-        except ValueError as e:
-            print(f"Error: {e}", file=sys.stderr)
-            sys.exit(1)
-
-        if args.fix:
-            import warnings
-
-            fix_cmd = "skillsaw fix --llm" if args.use_llm else "skillsaw fix"
-            msg = (
-                f"`skillsaw lint --fix` is deprecated and will be removed in 1.0. "
-                f"Use `{fix_cmd}` instead."
-            )
-            warnings.warn(msg, DeprecationWarning, stacklevel=1)
-            print(f"Warning: {msg}", file=sys.stderr)
-            violations, fixes = linter.fix()
-            applied = linter.apply_fixes(fixes)
-            if applied and args.fmt == "text":
-                print(f"Fixed {len(applied)} issue(s):")
-                for fix in applied:
-                    print(f"  ✓ [{fix.file_path}] {fix.description}")
-                print()
-            suggested = [f for f in fixes if f not in applied]
-            if suggested and args.fmt == "text":
-                print(f"Suggested fixes ({len(suggested)} — review before applying):")
-                for fix in suggested:
-                    print(f"  ? [{fix.file_path}] {fix.description}")
-                print()
-            if args.use_llm:
-                _run_llm_fix_inline(args, linter, config)
-        else:
-            rule_progress = _RuleProgress(args)
-            try:
-                violations = linter.run(progress=rule_progress)
-            finally:
-                rule_progress.clear()
-
-        all_violations.extend(violations)
-        all_rules.extend(linter.rules)
-        baseline_suppressed += linter.baseline_suppressed_count
-
-        if baseline and args.fmt == "text":
-            stale = linter.stale_baseline_entries
-            if stale:
-                print(
-                    f"Baseline: {len(stale)} stale"
-                    f" {'entry' if len(stale) == 1 else 'entries'}"
-                    f" (violations resolved since baseline was set)"
-                )
-                if args.verbose:
-                    for entry in stale:
-                        location = f" [{entry.file_path}]" if entry.file_path else ""
-                        print(f"  - {entry.rule_id}{location}: {entry.message}")
-                print("  Run `skillsaw baseline` to update.\n")
-
-    merged_context = _build_merged_context(contexts)
-    unique_rules = _dedup_rules(all_rules)
-    lint_duration = time.perf_counter() - lint_started
-
-    from .grade import compute_grade
-
-    content_tokens = sum(
-        block.estimate_tokens() for ctx in contexts for block in ctx.lint_tree.content_blocks()
-    )
-    grade = compute_grade(all_violations, content_tokens)
-
-    stdout_output = format_report(
-        args.fmt,
-        all_violations,
-        merged_context,
-        unique_rules,
-        cli_version,
-        verbose=args.verbose,
-        baseline_suppressed=baseline_suppressed,
-        duration=lint_duration,
-        grade=grade,
-    )
-    print(stdout_output)
-
-    report_cache = {}
-    for output_path, fmt in output_formats.items():
-        if fmt not in report_cache:
-            report_cache[fmt] = format_report(
-                fmt,
-                all_violations,
-                merged_context,
-                unique_rules,
-                cli_version,
-                verbose=args.verbose,
-                baseline_suppressed=baseline_suppressed,
-                duration=lint_duration,
-                grade=grade,
-            )
-        out_path = Path(output_path)
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        out_path.write_text(report_cache[fmt], encoding="utf-8")
-
-    errors, warnings_count, info = get_counts(all_violations)
-
-    if missing_count > 0 or errors > 0:
-        sys.exit(1)
-    elif config.strict and warnings_count > 0:
-        sys.exit(1)
-    else:
-        sys.exit(0)
-
-
-def _ansi_colors():
-    no_color = "NO_COLOR" in os.environ
-    return {
-        "bold": "" if no_color else "\033[1m",
-        "dim": "" if no_color else "\033[2m",
-        "green": "" if no_color else "\033[92m",
-        "red": "" if no_color else "\033[91m",
-        "yellow": "" if no_color else "\033[93m",
-        "cyan": "" if no_color else "\033[96m",
-        "reset": "" if no_color else "\033[0m",
-        "no_color": no_color,
-    }
-
-
-_DEFAULT_PATCH_NAME = ".skillsaw-llm-patch.diff"
-
-
-def _resolve_patch_path(args, root_path: Path) -> Path:
-    if getattr(args, "patch_file", None):
-        return args.patch_file.resolve()
-    return root_path.resolve() / _DEFAULT_PATCH_NAME
-
-
-def _save_llm_patch(diffs, patch_path: Path) -> None:
-    combined = ""
-    for diff_text in diffs.values():
-        combined += diff_text
-        if not combined.endswith("\n"):
-            combined += "\n"
-    patch_path.write_text(combined, encoding="utf-8")
-
-
-def _apply_llm_patch(patch_path: Path, root_path: Path) -> None:
-    import subprocess
-
-    c = _ansi_colors()
-    if not patch_path.exists():
-        print(f"Error: Patch file not found: {patch_path}", file=sys.stderr)
-        sys.exit(1)
-
-    patch_content = patch_path.read_text(encoding="utf-8")
-    if not patch_content.strip():
-        print(f"Error: Patch file is empty: {patch_path}", file=sys.stderr)
-        sys.exit(1)
-
-    result = subprocess.run(
-        ["git", "apply", "--check", str(patch_path)],
-        cwd=str(root_path),
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        print(
-            f"Error: Patch does not apply cleanly:\n{result.stderr}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    result = subprocess.run(
-        ["git", "apply", str(patch_path)],
-        cwd=str(root_path),
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        print(f"Error: Failed to apply patch:\n{result.stderr}", file=sys.stderr)
-        sys.exit(1)
-
-    print(f"{c['green']}Patch applied successfully from {patch_path}{c['reset']}")
-    patch_path.unlink()
-    print(f"{c['dim']}Patch file removed.{c['reset']}")
-
-
-def _require_llm_provider(config):
-    if not config.llm.model:
-        print(
-            "Error: No model configured. Set llm.model in your config file,"
-            " pass --model, or set SKILLSAW_MODEL.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    try:
-        from .llm._litellm import LiteLLMProvider
-    except ImportError:
-        print(
-            "Error: LLM features require litellm. Install with: pip install skillsaw[llm]",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    logging.basicConfig(level=logging.WARNING, format="%(name)s: %(message)s", stream=sys.stderr)
-    return LiteLLMProvider()
-
-
-def _print_colored_diff(diffs, c, header=None, separator=False):
-    if not diffs:
-        return
-    if header:
-        print(f"\n{c['bold']}{header}{c['reset']}")
-    if separator:
-        print(f"{c['dim']}{'─' * 60}{c['reset']}")
-    for diff_text in diffs.values():
-        for line in diff_text.splitlines():
-            if line.startswith("+") and not line.startswith("+++"):
-                print(f"{c['green']}{line}{c['reset']}")
-            elif line.startswith("-") and not line.startswith("---"):
-                print(f"{c['red']}{line}{c['reset']}")
-            elif line.startswith("@@"):
-                print(f"{c['cyan']}{line}{c['reset']}")
-            else:
-                print(line)
-    if separator:
-        print(f"{c['dim']}{'─' * 60}{c['reset']}")
-
-
-def _print_token_usage(usage, c, indent=""):
-    total_tokens = usage.prompt_tokens + usage.completion_tokens
-    if total_tokens:
-        print(f"{indent}{c['dim']}~{total_tokens:,} tokens{c['reset']}")
-
-
-def _run_llm_fix_inline(args, linter, config):
-    """Handle --fix --llm from the lint subcommand."""
-    c = _ansi_colors()
-    provider = _require_llm_provider(config)
-    dry_run = getattr(args, "dry_run", False)
-
-    import time
-
-    start_time = time.monotonic()
-
-    def _on_event(event_type, **kw):
-        if event_type == "progress":
-            elapsed = time.monotonic() - start_time
-            print(
-                f"{c['dim']}  [{kw['completed']}/{kw['file_count']} files,"
-                f" {int(elapsed)}s]{c['reset']}",
-                file=sys.stderr,
-            )
-        elif event_type == "file_start":
-            print(
-                f"  {c['bold']}{kw['rel_path']}{c['reset']}"
-                f"  {c['dim']}{kw['num_violations']} violation(s){c['reset']}",
-                file=sys.stderr,
-            )
-        elif event_type == "file_done":
-            remaining = kw.get("remaining", 0)
-            changed = kw.get("changed", False)
-            if not changed:
-                print(f"    {c['yellow']}no changes{c['reset']}", file=sys.stderr)
-            elif remaining == 0:
-                print(
-                    f"    {c['green']}✓ all {kw['num_violations']}"
-                    f" violation(s) fixed{c['reset']}",
-                    file=sys.stderr,
-                )
-            else:
-                fixed = kw["num_violations"] - remaining
-                print(
-                    f"    {c['red']}{fixed} fixed, {remaining} failed{c['reset']}",
-                    file=sys.stderr,
-                )
-
-    result = linter.llm_fix(
-        provider,
-        callback=_on_event,
-        min_severity=Severity.WARNING,
-        max_workers=config.llm.max_workers,
-        dry_run=dry_run,
-    )
-
-    _print_colored_diff(result.diffs, c, header="LLM Changes")
-
-    if dry_run:
-        print(f"\n{c['yellow']}dry-run — no files were modified{c['reset']}")
-        if result.diffs:
-            patch_path = _resolve_patch_path(args, linter.context.root_path)
-            _save_llm_patch(result.diffs, patch_path)
-            print(f"{c['cyan']}Patch saved to {patch_path}{c['reset']}")
-            print(f"\n{c['bold']}To apply:{c['reset']}" f" skillsaw fix --apply-patch")
-
-    _print_token_usage(result.total_usage, c)
-
-
-def _run_fix(args):
-    # Unlike lint, fix mutates files — fail fast on any missing path
-    # before touching anything, referencing exactly what the user typed
-    missing = [p for p in args.path if not p.exists()]
-    for p in missing:
-        print(f"Error: Path not found: {p}", file=sys.stderr)
-    if missing:
-        sys.exit(1)
-
-    paths = _resolve_lint_paths(args.path)
-
-    if getattr(args, "apply_patch", False):
-        if len(paths) > 1:
-            print("Error: --apply-patch accepts a single path", file=sys.stderr)
-            sys.exit(1)
-        root_path = paths[0]
-        patch_path = _resolve_patch_path(args, root_path)
-        _apply_llm_patch(patch_path, root_path)
-        sys.exit(0)
-
-    if args.config:
-        config_path = args.config
-        if not config_path.exists():
-            print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        config_path = find_config(paths[0])
-
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-
-    rule_ids = set(args.rule_ids) if args.rule_ids else None
-    skip_rule_ids = set(args.skip_rule_ids) if args.skip_rule_ids else None
-    if rule_ids and skip_rule_ids:
-        print("Error: --rule and --skip-rule cannot be combined", file=sys.stderr)
-        sys.exit(1)
-
-    if not args.use_llm:
-        import difflib
-
-        dry_run = getattr(args, "dry_run", False)
-        confidence = AutofixConfidence.SUGGEST if args.suggest else AutofixConfidence.SAFE
-
-        # (fix, repo root) pairs — the root relativizes paths in dry-run diffs
-        applied = []
-        suggested = []
-        for fix_path in paths:
-            context = RepositoryContext(fix_path)
-            try:
-                linter = Linter(
-                    context,
-                    config,
-                    rule_ids=rule_ids,
-                    skip_rule_ids=skip_rule_ids,
-                    no_custom_rules=args.no_custom_rules,
-                )
-            except ValueError as e:
-                print(f"Error: {e}", file=sys.stderr)
-                sys.exit(1)
-
-            rule_progress = _RuleProgress(args)
-            try:
-                path_applied, path_suggested = linter.fix_and_apply(
-                    confidence, dry_run=dry_run, progress=rule_progress
-                )
-            finally:
-                rule_progress.clear()
-
-            if not dry_run and any(f.rule_id == "agentskill-name" for f in path_applied):
-                context = RepositoryContext(fix_path)
-                linter = Linter(
-                    context,
-                    config,
-                    rule_ids=rule_ids,
-                    skip_rule_ids=skip_rule_ids,
-                    no_custom_rules=args.no_custom_rules,
-                )
-                rename_applied, rename_suggested = linter.fix_and_apply(confidence)
-                path_applied.extend(rename_applied)
-                path_suggested.extend(rename_suggested)
-
-            applied.extend((f, context.root_path) for f in path_applied)
-            suggested.extend(path_suggested)
-
-        c = _ansi_colors()
-
-        if applied:
-            label = "Would fix" if dry_run else "Fixed"
-            print(f"{label} {len(applied)} issue(s):")
-            for fix, root in applied:
-                print(f"  {c['bold']}✓ [{fix.file_path}] {fix.description}{c['reset']}")
-                if dry_run and fix.original_content != fix.fixed_content:
-                    try:
-                        rel = fix.file_path.relative_to(root)
-                    except ValueError:
-                        rel = fix.file_path
-                    diff_lines = difflib.unified_diff(
-                        fix.original_content.splitlines(keepends=True),
-                        fix.fixed_content.splitlines(keepends=True),
-                        fromfile=f"a/{rel}",
-                        tofile=f"b/{rel}",
-                    )
-                    for line in diff_lines:
-                        line = line.rstrip("\n")
-                        if line.startswith("+") and not line.startswith("+++"):
-                            print(f"      {c['green']}{line}{c['reset']}")
-                        elif line.startswith("-") and not line.startswith("---"):
-                            print(f"      {c['red']}{line}{c['reset']}")
-                        elif line.startswith("@@"):
-                            print(f"      {c['cyan']}{line}{c['reset']}")
-                        else:
-                            print(f"      {line}")
-                    print(f"      {c['dim']}{'─' * 40}{c['reset']}")
-        else:
-            print("No auto-fixable violations found.")
-
-        if suggested:
-            print(f"\nSuggested fixes ({len(suggested)} — review before applying):")
-            for fix in suggested:
-                print(f"  ? [{fix.file_path}] {fix.description}")
-            print("\nRun `skillsaw fix --suggest` to apply suggested fixes.")
-            print("Run `skillsaw fix --suggest --dry-run` to preview changes.")
-
-        if dry_run and applied:
-            print(f"\n{c['yellow']}dry-run — no files were modified{c['reset']}")
-
-        sys.exit(0)
-
-    # LLM fix runs an interactive session against one repository
-    if len(paths) > 1:
-        print("Error: --llm accepts a single path", file=sys.stderr)
-        sys.exit(1)
-
-    context = RepositoryContext(paths[0])
-    try:
-        linter = Linter(
-            context,
-            config,
-            rule_ids=rule_ids,
-            skip_rule_ids=skip_rule_ids,
-            no_custom_rules=args.no_custom_rules,
-        )
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    if args.model:
-        config.llm.model = args.model
-    if args.max_iterations is not None:
-        if args.max_iterations < 1:
-            print(
-                f"Error: --max-iterations must be >= 1, got {args.max_iterations}",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        config.llm.max_iterations = args.max_iterations
-
-    c = _ansi_colors()
-    provider = _require_llm_provider(config)
-
-    is_tty = sys.stdout.isatty()
-    term_size = shutil.get_terminal_size((80, 24))
-    term_width = term_size.columns
-    term_rows = term_size.lines
-
-    violations = linter.run()
-    llm_rules = {r.rule_id: r for r in linter.rules if r.llm_fix_prompt is not None}
-    min_severity = Severity.INFO if args.all else Severity.WARNING
-    llm_violations = [
-        v
-        for v in violations
-        if v.rule_id in llm_rules
-        and Linter._SEVERITY_ORDER.get(v.severity, 99) <= Linter._SEVERITY_ORDER[min_severity]
-    ]
-
-    files_with_violations = set()
-    for v in llm_violations:
-        if v.file_path:
-            files_with_violations.add(v.file_path.resolve())
-
-    print(f"\n{c['bold']}skillsaw fix{c['reset']} {c['dim']}({config.llm.model}){c['reset']}")
-    print(f"{len(llm_violations)} violation(s) across " f"{len(files_with_violations)} file(s)\n")
-
-    if not llm_violations:
-        print(f"{c['green']}No fixable violations found.{c['reset']}")
-        sys.exit(0)
-
-    def _progress_bar(current, total, width=20):
-        filled = int(width * current / total) if total else 0
-        bar = "█" * filled + "░" * (width - filled)
-        return f"{c['dim']}[{c['reset']}{c['cyan']}{bar}{c['reset']}{c['dim']}]{c['reset']}"
-
-    def _setup_scroll_region():
-        if not is_tty or c["no_color"]:
-            return
-        sys.stdout.write(f"\033[1;{term_rows - 1}r")
-        sys.stdout.write(f"\033[2J")
-        sys.stdout.write(f"\033[1;1H")
-        sys.stdout.flush()
-
-    def _update_status_bar(text):
-        if not is_tty or c["no_color"]:
-            return
-        padded = text[:term_width].ljust(term_width)
-        sys.stdout.write(f"\033[s\033[{term_rows};1H\033[2K{padded}\033[u")
-        sys.stdout.flush()
-
-    def _teardown_scroll_region():
-        if not is_tty or c["no_color"]:
-            return
-        sys.stdout.write(f"\033[1;{term_rows}r")
-        sys.stdout.write(f"\033[{term_rows};1H\033[2K")
-        sys.stdout.flush()
-
-    def _on_event(event_type, **kw):
-        if event_type == "file_start":
-            print(
-                f"{c['bold']}{kw['rel_path']}{c['reset']}"
-                f"  {c['dim']}{kw['num_violations']} violation(s):"
-                f" {', '.join(kw['rule_ids'])}{c['reset']}"
-            )
-        elif event_type == "iteration":
-            tag = f"{c['dim']}[{kw['rel_path']}]{c['reset']} "
-            print(
-                f"  {tag}{c['dim']}iteration"
-                f" {kw['iteration']}/{kw['max_iterations']}{c['reset']}"
-            )
-        elif event_type == "tool_call":
-            tool_args = kw.get("arguments", {})
-            arg_summary = ""
-            if "path" in tool_args:
-                arg_summary = str(tool_args["path"])
-            elif tool_args:
-                first_key = next(iter(tool_args))
-                val = str(tool_args[first_key])
-                if len(val) > 40:
-                    val = val[:37] + "..."
-                arg_summary = val
-            tag = f"{c['dim']}[{kw['rel_path']}]{c['reset']} "
-            print(f"  {tag}{kw['name']}({arg_summary})")
-        elif event_type == "retry":
-            tag = f"{c['dim']}[{kw['rel_path']}]{c['reset']} "
-            print(
-                f"  {tag}{c['yellow']}{kw['remaining']} violation(s)"
-                f" remain, retrying...{c['reset']}"
-            )
-        elif event_type == "file_done":
-            remaining = kw.get("remaining", 0)
-            changed = kw.get("changed", False)
-            tag = f"{c['dim']}[{kw['rel_path']}]{c['reset']} "
-            if not changed:
-                print(f"  {tag}{c['yellow']}no changes{c['reset']}")
-            elif remaining == 0:
-                print(
-                    f"  {tag}{c['green']}✓ all {kw['num_violations']}"
-                    f" violation(s) fixed{c['reset']}"
-                )
-            else:
-                fixed = kw["num_violations"] - remaining
-                print(f"  {tag}{c['red']}{fixed} fixed," f" {remaining} failed{c['reset']}")
-            print()
-
-    max_workers = args.workers or config.llm.max_workers
-
-    import time
-
-    start_time = time.monotonic()
-
-    def _on_event_with_timer(event_type, **kw):
-        if event_type == "progress":
-            elapsed = time.monotonic() - start_time
-            elapsed_str = f"{int(elapsed)}s"
-            eta_str = ""
-            if 0 < kw["completed"] < kw["file_count"]:
-                rate = elapsed / kw["completed"]
-                remaining = rate * (kw["file_count"] - kw["completed"])
-                eta_str = f" ETA {int(remaining)}s"
-            bar = _progress_bar(kw["completed"], kw["file_count"])
-            status = f" {kw['completed']}/{kw['file_count']} files {elapsed_str}{eta_str}"
-            _update_status_bar(f" {bar}{status}")
-            return
-        _on_event(event_type, **kw)
-
-    dry_run = getattr(args, "dry_run", False)
-
-    _setup_scroll_region()
-    try:
-        result = linter.llm_fix(
-            provider,
-            callback=_on_event_with_timer,
-            min_severity=min_severity,
-            max_workers=max_workers,
-            dry_run=dry_run,
-        )
-    finally:
-        _teardown_scroll_region()
-
-    if not result.success:
-        if dry_run:
-            print(f"\n{c['yellow']}LLM fix would not improve violations.{c['reset']}")
-            sys.exit(0)
-        print(f"\n{c['red']}LLM fix did not improve violations" f" — changes reverted.{c['reset']}")
-        sys.exit(1)
-
-    if not result.diffs and not result.files_modified:
-        print(f"{c['green']}No fixable violations found.{c['reset']}")
-        sys.exit(0)
-
-    label = "Dry-run results" if dry_run else "Results"
-    print(f"\n{c['bold']}{label}{c['reset']}")
-    print(
-        f"  {c['green']}{result.violations_fixed} fixed{c['reset']}"
-        f" {c['dim']}of {result.violations_before}{c['reset']}"
-    )
-    if result.violations_after > 0:
-        print(f"  {c['yellow']}{result.violations_after} remaining{c['reset']}")
-    if dry_run:
-        print(f"  {c['yellow']}dry-run — no files were modified{c['reset']}")
-        if result.diffs:
-            patch_path = _resolve_patch_path(args, context.root_path)
-            _save_llm_patch(result.diffs, patch_path)
-            print(f"  {c['cyan']}Patch saved to {patch_path}{c['reset']}")
-            print(f"\n  {c['bold']}To apply:{c['reset']}" f" skillsaw fix --apply-patch")
-    else:
-        print(f"  {len(result.files_modified)} file(s) modified")
-
-    _print_token_usage(result.total_usage, c, indent="  ")
-    _print_colored_diff(result.diffs, c, header="Changes", separator=True)
-
-    sys.exit(0)
-
-
-def _run_baseline(args):
-    if not args.path.exists():
-        print(f"Error: Path not found: {args.path}", file=sys.stderr)
-        sys.exit(1)
-
-    context = RepositoryContext(args.path)
-
-    if args.config:
-        config_path = args.config
-        if not config_path.exists():
-            print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        config_path = find_config(args.path)
-
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-
-    try:
-        linter = Linter(context, config)
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    violations = [v for v in linter.run() if v.severity != Severity.INFO]
-
-    from .baseline import build_baseline, save_baseline, BASELINE_FILENAME
-
-    cli_version = _get_version()
-    baseline_modes = {r.rule_id: r.baseline_mode for r in linter.rules if r.baseline_mode}
-
-    if config_path:
-        output_path = config_path.parent / BASELINE_FILENAME
-    else:
-        output_path = args.path / BASELINE_FILENAME
-
-    # Fingerprint paths relative to the directory the baseline is written to,
-    # so the baseline still matches when lint is later run from a subdirectory.
-    baseline = build_baseline(violations, output_path.resolve().parent, cli_version, baseline_modes)
-
-    save_baseline(output_path, baseline)
-    print(f"Baselined {len(baseline.violations)} violation(s) to {output_path}")
-    sys.exit(0)
-
-
-_BADGE_FILENAME = ".skillsaw-badge.json"
-
-
-def _github_raw_url(root_path: Path, badge_path: Path):
-    """Best-effort raw.githubusercontent.com URL for the badge file.
-
-    Returns None when the repo has no recognizable GitHub remote.
-    """
-    import re
-    import subprocess
-
-    def _git(*argv):
-        try:
-            result = subprocess.run(
-                ["git", "-C", str(root_path), *argv],
-                capture_output=True,
-                text=True,
-            )
-        except OSError:
-            return None
-        return result.stdout.strip() if result.returncode == 0 else None
-
-    remote = _git("config", "--get", "remote.origin.url")
-    if not remote:
-        return None
-    m = re.search(r"github\.com[:/]([^/]+)/(.+?)(?:\.git)?/?$", remote)
-    if not m:
-        return None
-    owner, repo = m.group(1), m.group(2)
-
-    branch = _git("symbolic-ref", "--short", "refs/remotes/origin/HEAD")
-    if branch and branch.startswith("origin/"):
-        branch = branch[len("origin/") :]
-    if not branch:
-        branch = _git("rev-parse", "--abbrev-ref", "HEAD") or "main"
-
-    try:
-        rel = badge_path.resolve().relative_to(root_path.resolve())
-    except ValueError:
-        rel = Path(badge_path.name)
-    return f"https://raw.githubusercontent.com/{owner}/{repo}/{branch}/{rel.as_posix()}"
-
-
-def _run_badge(args):
-    import json
-    from urllib.parse import quote
-
-    if not args.path.exists():
-        print(f"Error: Path not found: {args.path}", file=sys.stderr)
-        sys.exit(1)
-
-    context = RepositoryContext(args.path)
-
-    if args.config:
-        config_path = args.config
-        if not config_path.exists():
-            print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        config_path = find_config(args.path)
-
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-
-    # No baseline on purpose — the published grade must reflect every
-    # violation, not just the ones newer than the baseline.
-    try:
-        linter = Linter(context, config)
-    except ValueError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    rule_progress = _RuleProgress(args)
-    try:
-        violations = linter.run(progress=rule_progress)
-    finally:
-        rule_progress.clear()
-
-    from .grade import compute_grade
-
-    content_tokens = sum(b.estimate_tokens() for b in context.lint_tree.content_blocks())
-    grade = compute_grade(violations, content_tokens)
-
-    badge_path = args.output or (context.root_path / _BADGE_FILENAME)
-    badge_path.parent.mkdir(parents=True, exist_ok=True)
-    badge_path.write_text(json.dumps(grade.badge_json(), indent=2) + "\n", encoding="utf-8")
-
-    c = _ansi_colors()
-    grade_color = (
-        c["green"]
-        if grade.letter[0] in "AB"
-        else (c["yellow"] if grade.letter[0] == "C" else c["red"])
-    )
-    print(f"Grade: {grade_color}{c['bold']}{grade.letter}{c['reset']}")
-    print(
-        f"  {grade.errors} error(s), {grade.warnings} warning(s), {grade.info} info"
-        f" across ~{grade.content_tokens:,} content tokens"
-        f" ({grade.density:.2f} weighted violations per 10k tokens)"
-    )
-    print(f"\nBadge written to {badge_path}")
-
-    raw_url = _github_raw_url(context.root_path, badge_path)
-    if raw_url is None:
-        raw_url = "<RAW_URL_TO_YOUR_BADGE_JSON>"
-        print(
-            "\nNo GitHub remote detected — replace the URL placeholder after"
-            " publishing the badge file somewhere shields.io can fetch it."
-        )
-    from .grade import logo_data_uri
-
-    encoded = quote(raw_url, safe="")
-    label = "skillsaw"
-    logo = quote(logo_data_uri(), safe="")
-
-    print(f"\n{c['bold']}Markdown for your README:{c['reset']}")
-    print(f"\n  Dynamic JSON badge (https://shields.io/badges/dynamic-json-badge):")
-    print(
-        f"  [![skillsaw grade](https://img.shields.io/badge/dynamic/json"
-        f"?url={encoded}&query=%24.message&label={label}&color={grade.color}"
-        f"&logo={logo})](https://skillsaw.org/)"
-    )
-    print(f"\n  Endpoint badge (color updates with the grade automatically):")
-    print(
-        f"  [![skillsaw grade](https://img.shields.io/endpoint?url={encoded})]"
-        f"(https://skillsaw.org/)"
-    )
-    print(
-        f"\nCommit {badge_path.name} and regenerate it (e.g. in CI) whenever"
-        " the repository changes."
-    )
-    sys.exit(0)
-
-
-def _run_init(args):
-    config_path = args.path / ".skillsaw.yaml"
-    if config_path.exists():
-        print(f"Config file already exists: {config_path}")
-        sys.exit(1)
-
-    config = LinterConfig.for_init()
-    config.save(config_path)
-    print(f"Created default config: {config_path}")
-    print("Run `skillsaw baseline` to accept existing violations.")
-    sys.exit(0)
-
-
-def _run_list_rules():
-    from .rules.builtin import BUILTIN_RULES
-
-    print("Available builtin rules:\n")
-    for rule_class in BUILTIN_RULES:
-        rule = rule_class()
-        fix_types = []
-        if rule.supports_autofix:
-            fix_types.append("auto")
-        if rule.llm_fix_prompt is not None:
-            fix_types.append("llm")
-        fix_label = ", ".join(fix_types) if fix_types else "none"
-        print(f"  {rule.rule_id}")
-        print(f"    {rule.description}")
-        print(f"    Default severity: {rule.default_severity().value}")
-        print(f"    Autofix: {fix_label}")
-        print()
-    sys.exit(0)
-
-
-def _run_explain(args):
-    from .rules.builtin import BUILTIN_RULES
-    from .rule_docs import load_rule_docs, rule_doc_url
-
-    if not args.path.exists():
-        print(f"Error: Path not found: {args.path}", file=sys.stderr)
-        sys.exit(1)
-    if args.config and not args.config.exists():
-        print(f"Error: Config file not found: {args.config}", file=sys.stderr)
-        sys.exit(1)
-
-    rule_class = None
-    known_ids = []
-    for candidate in BUILTIN_RULES:
-        rule = candidate()
-        known_ids.append(rule.rule_id)
-        if rule.rule_id == args.rule_id:
-            rule_class = candidate
-            break
-
-    if rule_class is None:
-        print(f"Error: Unknown rule '{args.rule_id}'", file=sys.stderr)
-        import difflib
-
-        close = difflib.get_close_matches(args.rule_id, known_ids, n=3)
-        if close:
-            print(f"Did you mean: {', '.join(close)}?", file=sys.stderr)
-        print("Run `skillsaw list-rules` to see all rules.", file=sys.stderr)
-        sys.exit(1)
-
-    c = _ansi_colors()
-    defaults = LinterConfig.default()
-    default_rule = rule_class(defaults.get_rule_config(args.rule_id))
-    default_enabled = defaults.get_rule_config(args.rule_id).get("enabled", True)
-
-    fix_types = []
-    if default_rule.supports_autofix:
-        fix_types.append("auto")
-    autofix_label = ", ".join(fix_types) if fix_types else "none"
-    llm_fix_label = "yes" if default_rule.llm_fix_prompt is not None else "no"
-
-    header_meta = (
-        f"{default_rule.severity.value}, autofix: {autofix_label}, "
-        f"llm-fix: {llm_fix_label}, since {default_rule.since}"
-    )
-    print(f"{c['bold']}{args.rule_id}{c['reset']} ({header_meta})")
-    print()
-    print(default_rule.description)
-
-    long_docs = load_rule_docs(args.rule_id)
-    if long_docs:
-        print()
-        print(long_docs)
-
-    if default_rule.repo_types:
-        repo_types_str = ", ".join(sorted(t.value for t in default_rule.repo_types))
-        print()
-        print(f"{c['bold']}Applies to repo types:{c['reset']} {repo_types_str}")
-
-    print()
-    print(f"{c['bold']}Configuration{c['reset']} (.skillsaw.yaml):")
-    print("  rules:")
-    print(f"    {args.rule_id}:")
-    enabled_str = LinterConfig._yaml_value(default_enabled)
-    print(f"      enabled: {enabled_str}        # true | false | auto")
-    print(f"      severity: {default_rule.severity.value}     " f"# error | warning | info")
-    for param_name, param_info in default_rule.config_schema.items():
-        default_val = LinterConfig._yaml_value(param_info.get("default"), indent=8)
-        desc = param_info.get("description", "")
-        if default_val.startswith("\n"):
-            print(f"      {param_name}:  # {desc}{default_val}")
-        else:
-            print(f"      {param_name}: {default_val}  # {desc}")
-
-    # Effective config in the target repository — doubles as a debugging
-    # tool for "why didn't this rule fire?"
-    context = RepositoryContext(args.path)
-    config_path = args.config if args.config else find_config(args.path)
-    config = None
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    if config is None:
-        config = LinterConfig.default()
-        config_label = "builtin defaults"
-    else:
-        config_label = str(config_path)
-
-    enabled, reason = config.rule_enabled_reason(
-        args.rule_id,
-        context,
-        default_rule.repo_types,
-        default_rule.formats,
-        since_version=default_rule.since,
-    )
-    try:
-        effective_rule = rule_class(config.get_rule_config(args.rule_id))
-        effective_severity = effective_rule.severity.value
-    except ValueError:
-        effective_severity = "(invalid severity in config)"
-
-    state = f"{c['green']}enabled{c['reset']}" if enabled else f"{c['red']}disabled{c['reset']}"
-    print()
-    print(f"{c['bold']}Effective in {args.path.resolve()}{c['reset']} ({config_label}):")
-    print(f"  {state} — {reason}")
-    if enabled:
-        print(f"  severity: {effective_severity}")
-
-    print()
-    print(f"{c['bold']}Docs:{c['reset']} {rule_doc_url(args.rule_id)}")
-    sys.exit(0)
-
-
-def _run_docs(args):
-    if not args.path.exists():
-        print(f"Error: Path not found: {args.path}", file=sys.stderr)
-        sys.exit(1)
-
-    context = RepositoryContext(args.path)
-
-    if context.repo_type == RepositoryType.UNKNOWN:
-        print("Warning: Directory doesn't appear to be a recognized repository", file=sys.stderr)
-        print(
-            "Expected: .claude-plugin/plugin.json, plugins/ directory, or SKILL.md (agentskills.io)\n",
-            file=sys.stderr,
-        )
-
-    config_path = find_config(args.path)
-    if config_path:
-        try:
-            config = LinterConfig.from_file(config_path)
-        except ValueError as e:
-            print(f"Error loading config: {e}", file=sys.stderr)
-            sys.exit(1)
-        _emit_config_warnings(config)
-    else:
-        config = LinterConfig.default()
-    context.exclude_patterns = config.exclude_patterns
-    context.apply_excludes()
-
-    from .docs import extract_docs, render_html, render_markdown
-    from .docs.html_renderer import COLOR_THEMES
-
-    theme = args.theme
-    if theme and theme not in COLOR_THEMES:
-        print(
-            f"Error: Unknown theme '{theme}'. " f"Available: {', '.join(sorted(COLOR_THEMES))}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    docs_output = extract_docs(context, title=args.title)
-
-    if args.fmt == "html":
-        pages = render_html(docs_output, theme=theme)
-    else:
-        pages = render_markdown(docs_output)
-
-    output = args.output
-    if output and output.suffix in (".html", ".md"):
-        output.parent.mkdir(parents=True, exist_ok=True)
-        combined = "\n".join(pages.values()) if len(pages) > 1 else next(iter(pages.values()))
-        output.write_text(combined, encoding="utf-8")
-        print(f"Documentation written to {output}")
-    else:
-        output_dir = output or Path("skillsaw-docs")
-        output_dir.mkdir(parents=True, exist_ok=True)
-        for filename, content in pages.items():
-            (output_dir / filename).write_text(content, encoding="utf-8")
-        file_list = ", ".join(sorted(pages.keys()))
-        print(f"Documentation written to {output_dir}/ ({len(pages)} file(s): {file_list})")
-
-
-def claudelint_shim():
-    """Backward-compat entry point for the 'claudelint' command"""
-    print(
-        "WARNING: 'claudelint' has been renamed to 'skillsaw'. "
-        "Please update your scripts. The 'claudelint' command will be removed in a future release.",
-        file=sys.stderr,
-    )
-    main()
-
+from .cli import main, claudelint_shim
 
 if __name__ == "__main__":
     main()
